requirements.txt
# requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
torch 
transformers 
datasets 
scikit-learn
transformers[torch]
matplotlib
fastapi
uvicorn